* feat(ScriptFactory): Support custom engine class.

* feat:js connector basic part.

1. You can load JS in the registerCapabilities method, obtain the methods implemented by JS, and report them.
2. When calling the connection test, check whether the connectionTest method is implemented in the js. If it is implemented, call it. If it is not implemented, it will automatically return success. However, the test connection is not implemented in the info.
3. When calling discoverSchema() and  table_ count(), call JS discoverSchema and table_count().
4. When calling batchRead, call JS's batchRead.
5. When calling streamRead, call JS streamRead.

* feat:The post main API object needed to encapsulate the JS API.

* feat:Execute the corresponding APIInvoker with more types.

* feat: Java objects needed to encapsulate JS APIs .

1. BatchRead accept data object;
2. StreamRead accept data objec.

* feat: Encapsulation of paging iteration mode .

* feat(JavaScript-Connector): Adjust the partial structure of javascript connector .

1. The independent class package of the tapAPI, which contains the implementation of the iterate method. APIFactory/APIInvoker does not need to modify the code for JS, and independent wrapper classes are used to implement the difference part.
2. The method identification in the JS code is changed to judge whether the method object is a method after executing the code.
3. The dataTypes in spec.json are described by JS type.
4. Code related to APIFactory is moved from pdk api to api loader core.
5. Write a coding case in javascript-conector.

* feat(JS-CONNECTOR):change js engine to use.

* feat(JSConnector): Add JS processor configuration .

* fix(QuickApi): The http returned results are compatible with map and list transformations.

* fix(QuickAPI):json schema .

* feat(js-connector): Js supports paging and cache processing .

* feat(Js-Connector): Provide js tools for use .

* feat(JS-Connector): Provide cache processing and incremental cache processing, and complete js-connecotr with Coding as an example.

* fix(JS-Connector): Complete the coding data source case developed based on JavaScript, including incremental and full reading.

* fix(QuickApi): Perfect connection test .

* fix(JS-Connector): Code format .

* fix(JS-connector):json Schema.

* feat(js-connector):support command.

* feat(js-connector):json for array.

* Revert "Merge branch 'develop-v2.13' into feat-pdk-js-api"

This reverts commit aa476304843428338758e53d7d9acdfaf6f4d34e, reversing
changes made to 100102bfaa42f53e43466c0eb0377593758dd7bf.

* fix(JS-connector): Address where JSON API is loaded by default .

* feat(JS-connector):add a demo js connector.

* feat(JS-connector): Add a new js connector template.

+++ manager/tm/src/main/java/com/tapdata/tm/task/service/impl/TaskStartServiceImpl.java
@@ -71,6 +71,10 @@ public class TaskStartServiceImpl implements TaskStartService {
 
     @Override
     public boolean taskStartCheckLog(TaskDto taskDto, UserDetail userDetail) {
+        if (!TaskDto.SYNC_TYPE_MIGRATE.equals(taskDto.getSyncType())) {
+            return false;
+        }
+
         taskDagCheckLogService.removeAllByTaskId(taskDto.getId().toHexString());
 
         boolean saveNoPass = false;
@@ -79,6 +83,7 @@ public class TaskStartServiceImpl implements TaskStartService {
             Optional<TaskDagCheckLog> any = saveLogs.stream().filter(log -> Level.ERROR.equals(log.getGrade())).findAny();
             if (any.isPresent()) {
                 saveNoPass = true;
+                //taskService.updateStatus(taskDto.getId(), TaskDto.STATUS_EDIT);
             }
         }
 
@@ -88,6 +93,9 @@ public class TaskStartServiceImpl implements TaskStartService {
             Optional<TaskDagCheckLog> any = startLogs.stream().filter(log -> Level.ERROR.equals(log.getGrade())).findAny();
             if (any.isPresent()) {
                 startNoPass = true;
+                if (!saveNoPass) {
+                    //taskService.updateStatus(taskDto.getId(), TaskDto.STATUS_EDIT);
+                }
             }
         }
